# [WIP] Increase `work_mem` to speed up queries in Index page