--- a/app/models/aggregations/materialized_view.rb
+++ b/app/models/aggregations/materialized_view.rb
@@ -1,5 +1,5 @@
 class Aggregations::MaterializedView
   def self.prepare
-    DatabaseHelper.increase_work_mem_for_current_transaction
+    DatabaseHelper.increase_work_mem_for_current_transaction(megabytes: 500)
   end
 end