@pmanrubia 2019-02-13
Closed in favor of https://github.com/alphagov/content-performance-manager/pull/1149